requirements.txt
pandas
numpy 
cython
argparse
pysam
joblib
tqdm
lmfit
sklearn
pytest
